// src/app/models/firma.ts
export class Firma {
  frm_kod: string;
  frm_ad: string;
  frmg_ad: string;
  frm_ksad: string;
  frm_frmd_kod: string;
  frm_i3: string;
  frm_per_no: number = 0;
  frm_vergno: string;
  frm_vergda: string;
  frm_sehir: string;
  frm_muh_cari: string;
  frm_hsp_kod: string;
  frm_efat_sw: number | string;
  frm_ktgr_kod: string;
  frm_adres1: string;
  frm_yok: number;
  frm_stop: number;
  frm_mahal: string;
  frm_sokak: string;
  frm_binaadi: string;
  frm_binano: string;
  frm_kapino: string;
  frm_posta: string;
  frm_ulke: string;
  frm_ilce: string;
  frm_telefon: string;
  frm_fax: string;
  frm_yetkili: string;
  frm_web: string;
  frm_email: string;
  frm_cep_tel: string;
  frm_tcno: string;
  frm_tc_soyad: string;
  frm_odeme_vkn: string;
  frm_dvz_kod: string;
  frm_kur_tur: string;
  frm_ode_id: string;
  frm_iskonto: number;
  frm_nak_bedeli: number;
  frm_kur_fark: string;
  frm_ban_kod1: string;
  frm_hesno1: string;
  frm_iban1: string;
  frm_v2: string;
  frm_ahes: number;
  frm_ahes_asim: number;
  frm_kcek: number;
  frm_kcek_asim: number;
  frm_mcek_r1: string;
  frm_irs: number;
  frm_irs_asim: number;
  frm_sips_r1: string;
  frm_sipo: number;
  frm_isletme: string;
  frm_faiz: string;
  frm_pass: string;
  frm_cektest: string;
  frm_gs_id: number = 6;
  frm_amb_kod: string;
  frm_aciklama: string;
  frm_ent_kod2: string;
  frm_c1: string;
  frm_ahes_r1: string;
  frm_ahes_r2: string;
  frm_kcek_r1: string;
  frm_kcek_r2: string;
  frm_mcek_r2: string;
  frm_irs_r2: string;
  frm_irs_r1: string;
  frm_sips: number;
  frm_sips_r2: string;
  frm_sipo_r1: string;
  frm_sipo_asim: number;
  frm_sips_asim: number;
  frm_mcek_asim: number;
  frm_t3: number;
  frm_adres2: string;
  frm_adres3: string;
  frm_cadde: string;
  frm_ulke_kod: string;
  frm_tc_ad: string;
  frm_ode_tip: number;
  frm_ban_kod2: string;
  frm_hesno2: string;
  frm_iban2: string;
  frm_ban_kod3: string;
  frm_hesno3: string;
  frm_iban3: string;
  frm_ban_kod4: string;
  frm_hesno4: string;
  frm_iban4: string;
  frm_ban_kod5: string;
  frm_hesno5: string;
  frm_iban5: string;
  frm_v1: string;
  frm_mcek: number;
  frm_sipo_r2: string;
  frm_vade: string;
  frm_ent_kod1: string;
  frm_t1: number;
  frm_gs_id1: string;
  frm_gs_id2: string;
  frm_gs_id3: string;
  frm_gs_id4: string;
  frm_gs_id5: string;
  updt: Date;
  idt: Date;
  srk_no: number = Number(localStorage.getItem('srk_no'));
  Where?: { [key: string]: string };
frm_frmg_kod: any;
}